refactored find candidate response functionality to subclass
